requirements.txt
azure-functions
numpy==1.16.4
opencv_python==4.1.1.26
scikit_image==0.15.0
scipy==1.2.1
azure-storage-blob
